clean splitting functions
