+++ lib/model/bert.py
@@ -14,7 +14,6 @@ from transformers import AutoModel, AutoTokenizer, BatchEncoding
 
 from lib.entities.learning_curve import LossesSingleEpoch
 from lib.model.base import Model
-from lib.preprocessing.tokenize import tokenize_truncated
 
 load_dotenv()
 
@@ -127,7 +126,9 @@ class BertClassifier(Model):
 
     def _tokenize(self, texts: list[str]) -> BatchEncoding:
         """Transforms list of texts to list of tokens (truncated to 512 tokens)."""
-        tokens = tokenize_truncated(texts, self.tokenizer)
+        tokens = self.tokenizer.batch_encode_plus(
+            texts, max_length=512, padding=True, truncation=True, return_tensors="pt"
+        )
         return tokens
 
     def _train_single_epoch(self, dataloader: DataLoader, optimizer: Optimizer) -> tuple[float, float]:
